+++ tests/sentry/dynamic_sampling/tasks/helpers/test_boost_low_volume_transactions.py
